Don't panic when parsed data is too short
Some slcan devices seem to send unrecognized data sometimes. In this case, we should return an error instead of panicing. The present code should already cover the cases where the message kind is not understood (try_into() will fail with an error) and if `message_data` is too short for the found message type.